Change the way we transfer results and merge to the main repo
Instead of pulling the current version to the build apps and
creating a build with it as parent we now ignore that, sending
the entire build result to the repo. Then when we update the
master repo we create new commits based on the ones from the
builders, but with signatures and the correct parent commit.

This also has the advantage that we can create all the deltas
before we merge the import repo to the main repo.