Redo how we lock and structure the builds
We used to just lock on the project property, but now that things
can come from various places that is not reliable anymore, so we need
a new approach.

We create something called a build-id, which is normally just the
appid, but in the case of a versioned thing like a runtime we append
the known version with an @ (so "org.gnome.gedit" and
"org.kde.Sdk@5.9").  We then use this instead of the app id to name
the builder work dir.  However, we only want one thing using the work
dir at a time, so we need some locks.

Unfortunately the build-id is not known at the point of the build for
the entire thing, so we cannot just use it instead of project in the
current code.

So, instead we take the lock on the worker side, when actually
building this, because at this time the build id is known. This
additonally allows some more parallelization, in case we have multiple
changes to the same build-id in the queue.

This is nice, but it brings complications with handling the build
result, because that used to be scoped by the app id only. Instead
we now scope it by the build-number of the main build job.