Fix the github status reporter
The github status reporter assumes that if the project field has a
slash in it it is the github repo owner / repo name. This is not true
when we're using the forced-build-id hack. We fix this by not using
a slash then, which will force it to fall back to detecting this from
the repo url.