Generate static deltas on the builder
We now pull the latest official release from the flathub repo
into the builder once the build succeeded, then the build is
rebased on that using build-commit-from and deltas are
generated for this. The deltas are then included in the tarball
we upload.

When importing we rely on flatpak build-commit-from now
copying/rewriting the deltas, so we can just rsync any deltas from
import-repo to the real repo. In case anything went wrong with the
generated deltas and they don't apply anymore the existing call to
--generate-static-deltas will generate new ones.