Add keep-builds option
Unless this is set we clear the entire .flatpak-builder directory
after a build. This means all builds are from scratch, which makes
a lot of sense for flathub as when we start adding worker the chance
of getting the same worker is low, and leaving random builds on workers
is going to eventually run the workers out of space.