Use --timestamp=NOW in update-repo
This means we the timestamp is created on the same machine, and at the
time of the import, rather than the build machines. Doing it on the
build machines risk the timestamps for two immediate consecutive
builds getting a backwards-going timestamp if one arch builds faster on
the second build.