Use "soft purge" for invalidating summary file on CDN
As the summary file has a very low cache lifetime (one minute), if there is any temporary problem reaching the origin server, the CDN gives 503 errors for the summary after one minute. We've added "stale-if-error" to the Surrogate-Control header on the repo master, but this will have no effect if the summary file is explicitly purged from the CDN. Use the "soft purge" to mark the summary as needing refreshing immediately, but keep the stale object so it can be served in this situation. See https://docs.fastly.com/guides/purging/soft-purges.