General handling of non-stable branches
If no build is otherwise configured for this id, then if the
git branch is of the form "branch/FOO", then we take this to mean
that the flatpak branch should be FOO. And vice versa, when going
from buildname like "org.some.app/FOO" we assume you want the
git branch "branch/FOO".

This allows regular github repos to have multiple branches, for
example base apps.