Using repo manager
Instead of doing all the work on uploading to and importing into the
repo from buildbot we use a separate service for the repo manager.
You need to configure a repo manager url and a token for it.
Using this we create a repo manager build for each build, and
create a subset token to allow the build workers to upload
to that build. Then in the update-repo job we just tell
the repo manager to commit, publish and prune the build.

This change removes almost all buildbot config from this
config repo, and depends on using the forked buildbot
instead.