Add builds for qt5.15lts-21.08
Since Qt is not releasing anymore, we are still maintaining the one based on their last release. Since there's a new SDK release, we decided to create a new version of ours based on it.
The naming is a bit wonky for sure, but at least we get to use the last versions of our base.